# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: TradeUP API — Arabic trading education simulator with AI coach
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trading/stocks/{symbol}/history/{period}:
    get:
      operationId: getStockHistory
      tags: [trading]
      summary: Get price history for a stock
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: path
          required: true
          schema:
            type: string
            enum: [1d, 1w, 1m, 3m, 1y]
      responses:
        "200":
          description: Price history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PricePoint"
  /trading/stocks/{symbol}/analyze:
    post:
      operationId: analyzeStock
      tags: [trading]
      summary: Educational AI brief on a stock (not a recommendation)
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock analysis brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockAnalysis"
  /trading/orders:
    get:
      operationId: getOrders
      tags: [trading]
      summary: Get trade history
      responses:
        "200":
          description: Order history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [trading]
      summary: Place a buy or sell order (use confirmed=true after high-risk preview)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed with AI coaching feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResult"
  /trading/orders/preview:
    post:
      operationId: previewOrder
      tags: [trading]
      summary: Pre-trade AI coach review; may require confirmation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderPreviewInput"
      responses:
        "200":
          description: Preview with risk and coach feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPreview"
  /trading/coach/advice:
    post:
      operationId: getCoachAdvice
      tags: [trading]
      summary: Ask the AI trading coach a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachQuery"
      responses:
        "200":
          description: Coach response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachResponse"
  /trading/sessions/review:
    post:
      operationId: reviewTradingSession
      tags: [trading]
      summary: Calm post-session review aligned with coach philosophy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionReviewInput"
      responses:
        "200":
          description: Session review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionReview"
...
components:
  schemas:
...
    Portfolio:
      type: object
      required: [cashBalance, totalValue, totalGainLoss, totalGainLossPercent, holdings]
      properties:
        cashBalance:
          type: number
        totalValue:
          type: number
        totalGainLoss:
          type: number
        totalGainLossPercent:
          type: number
        dayGainLoss:
          type: number
        dayGainLossPercent:
          type: number
        holdings:
          type: array
          items:
            $ref: "#/components/schemas/Holding"
    Holding:
      type: object
      required: [symbol, name, shares, avgCost, currentPrice, totalValue, gainLoss, gainLossPercent]
      properties:
        symbol:
          type: string
        name:
          type: string
        shares:
          type: number
        avgCost:
          type: number
        currentPrice:
          type: number
        totalValue:
          type: number
        gainLoss:
          type: number
        gainLossPercent:
          type: number
        sector:
          type: string
...
    PricePoint:
      type: object
      required: [date, open, high, low, close, volume]
      properties:
        date:
          type: string
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: integer
    Order:
      type: object
      required: [id, symbol, type, shares, price, total, executedAt]
      properties:
        id:
          type: integer
        symbol:
          type: string
        type:
          type: string
          enum: [buy, sell]
        shares:
          type: number
        price:
          type: number
        total:
          type: number
        notes:
          type: ["string", "null"]
        executedAt:
          type: string
    OrderInput:
      type: object
      required: [symbol, type, shares]
      properties:
        symbol:
          type: string
        type:
          type: string
          enum: [buy, sell]
        shares:
          type: number
        notes:
          type: string
        confirmed:
          type: boolean
          description: Set true after acknowledging high-risk preview confirmation
    OrderPreviewInput:
      type: object
      required: [symbol, type, shares]
      properties:
        symbol:
          type: string
        type:
          type: string
          enum: [buy, sell]
        shares:
          type: number
    OrderPreview:
      type: object
      required: [riskLevel, requiresConfirm, coachFeedback, alignmentScore, philosophyNotes, questionsForUser]
      properties:
        riskLevel:
          type: string
          enum: [low, moderate, high, extreme]
        requiresConfirm:
          type: boolean
        warning:
          type: ["string", "null"]
        coachFeedback:
          type: string
        alignmentScore:
          type: number
        philosophyNotes:
          type: array
          items:
            type: string
        questionsForUser:
          type: array
          items:
            type: string
        estimatedTotal:
          type: number
        tradePercentOfPortfolio:
          type: number
    OrderResult:
      type: object
      required: [order, portfolio, coachFeedback, riskLevel]
      properties:
        order:
          $ref: "#/components/schemas/Order"
        portfolio:
          $ref: "#/components/schemas/Portfolio"
        coachFeedback:
          type: string
        riskLevel:
          type: string
          enum: [low, moderate, high, extreme]
        warning:
          type: ["string", "null"]
        xpGained:
          type: integer
    CoachHistoryMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
    CoachQuery:
      type: object
      required: [question]
      properties:
        question:
          type: string
          minLength: 1
        history:
          type: array
          items:
            $ref: "#/components/schemas/CoachHistoryMessage"
        coachId:
          type: string
          enum: [value, growth, risk, technical]
    CoachResponse:
      type: object
      required: [message, suggestions, personaEmotion]
      properties:
        message:
          type: string
        suggestions:
          type: array
          items:
            type: string
        personaEmotion:
          type: string
          enum: [thinking, excited, cautious, proud, teaching]
        learnMore:
          type: ["string", "null"]
        refs:
          type: array
          items:
            type: string
    SessionReviewInput:
      type: object
      properties:
        coachId:
          type: string
          enum: [value, growth, risk, technical]
        limit:
          type: integer
          default: 20
    SessionReview:
      type: object
      required: [summary, alignmentScore, whatWentWell, whatToImprove, nextPractice, personaEmotion]
      properties:
        summary:
          type: string
        alignmentScore:
          type: number
        whatWentWell:
          type: array
          items:
            type: string
        whatToImprove:
          type: array
          items:
            type: string
        nextPractice:
          type: string
        personaEmotion:
          type: string
          enum: [proud, thoughtful, cautious]
    StockAnalysis:
      type: object
      required: [symbol, brief, keyPoints, risks, educationalNote]
      properties:
        symbol:
          type: string
        brief:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: string
        educationalNote:
          type: string
...