# lib/api-zod/src/generated/types/coachHistoryMessage.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * TradeUP API — Arabic trading education simulator with AI coach
 * OpenAPI spec version: 0.2.0
 */
import type { CoachHistoryMessageRole } from './coachHistoryMessageRole';

export interface CoachHistoryMessage {
  role: CoachHistoryMessageRole;
  content: string;
}