# lib/api-zod/src/generated/types/coachQuery.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * TradeUP API — Arabic trading education simulator with AI coach
 * OpenAPI spec version: 0.2.0
 */
import type { CoachHistoryMessage } from './coachHistoryMessage';
import type { CoachQueryCoachId } from './coachQueryCoachId';

export interface CoachQuery {
  /** @minLength 1 */
  question: string;
  history?: CoachHistoryMessage[];
  coachId?: CoachQueryCoachId;
}