# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * TradeUP API — Arabic trading education simulator with AI coach
 * OpenAPI spec version: 0.2.0
 */
import type { OrderType } from './orderType';

export interface Order {
  id: number;
  symbol: string;
  type: OrderType;
  shares: number;
  price: number;
  total: number;
  /** @nullable */
  notes?: string | null;
  executedAt: string;
}